[HIL] Reverted float timestamp to uint32_t and converted it from seconds to milliseconds

+++ components/common/data/src/AccelerometerSensorData.hpp
@@ -17,7 +17,7 @@ public:
     float acceleration_z = 0.0f;
     
     // Metadata
-    float timestamp = 0;
+    uint32_t timestamp = 0;
 
     json toJSON() const override {
         json j;
