+++ components/sensors/Termoresistenze/src/Termoresistenze.hpp
@@ -25,7 +25,7 @@ public:
     double temperature;
 
     // Metadata
-    float timestamp;
+    uint32_t timestamp;
 
     json toJSON() const override {
         json j;
